requirements.txt
nibabel>=3.2
numpy>=1.2
pynrrd>=0.4.2
